+++ easy_store_boot/src/main/java/org/jeecgframework/boot/easy_store_boot/app/modules/api/vo/GoodsSearchResult.java
@@ -28,7 +28,7 @@ public class GoodsSearchResult {
 
     private Double costPrice;
 
-    private Integer stock;
+    private Double stock;
 
 
 }
