+++ easy_store_boot/src/main/java/org/jeecgframework/boot/easy_store_boot/app/modules/service/IAppSaleOrderItemService.java
@@ -22,7 +22,7 @@ public interface IAppSaleOrderItemService extends IService<AppSaleOrderItem> {
 
     void removeByUpdate(List<AppSaleOrderItem> updateList,Integer orderId);
 
-    Integer sumQuantityGoodsId(String goodsId);
+    Double sumQuantityGoodsId(String goodsId);
 
 
 }
